remove regelrett stuff from src code
