+++ .env.production
@@ -3,4 +3,3 @@ VITE_AUTHORITY=https://login.microsoftonline.com/7531b79e-fd42-4826-bff2-131d82c
 VITE_LOGIN_REDIRECT_URI=https://frisk-frontend.fly.dev
 VITE_BACKEND_URL=https://frisk-backend.fly.dev
 VITE_REGELRETT_FRONTEND_URL=https://regelrett-frontend-1024826672490.europe-north1.run.app
-VITE_REGELRETT_BACKEND_URL=http://regelrett.bekk.no:8080
