+++ .env.skip
@@ -3,4 +3,3 @@ VITE_AUTHORITY=https://login.microsoftonline.com/7f74c8a2-43ce-46b2-b0e8-b6306cb
 VITE_LOGIN_REDIRECT_URI=https://frisk.atgcp1-dev.kartverket-intern.cloud
 VITE_BACKEND_URL=https://api.frisk.atgcp1-dev.kartverket-intern.cloud
 VITE_REGELRETT_FRONTEND_URL=https://regelrett.atgcp1-dev.kartverket-intern.cloud
-VITE_REGELRETT_BACKEND_URL=http://regelrett-backend.regelrett-main:8080
